project-idea-generator/requirements.txt
requests==2.28.1
customtkinter==5.1.2
python-dotenv==1.0.0